fix(routing,seo): disputes URL is canonical + tighten robots.txt
Disputes route swap:
  Sidebar, hero CTAs and onboarding all link to /dashboard/disputes
  but the actual page lived at /dashboard/complaints — /disputes was
  a server-side redirect that flipped the URL bar back to the legacy
  path on every navigation. Users (rightly) expected the URL to stay
  on disputes.

  Now: /dashboard/disputes hosts the real page, /dashboard/complaints
  is the redirect target (preserves query params for deep-linked
  alerts/blogs/emails). 17 internal links to /complaints continue to
  work via the redirect; new code should link to /disputes directly.

robots.txt:
  Was Allow: / with no disallows, letting crawlers waste budget on
  /dashboard/, /auth/, /api/ surfaces that all return 401/403 and
  have zero SEO value. Now explicitly disallows those plus /preview/.
  Sitemap link preserved.
